Simplify source reading flow, improve file naming
